fix(notifications): resolve PR #392 review findings
- daily_plan: prefix-match the stale-nudge sweep so a snoozed plan.nudge
  (dedup `plan.nudge:{today}:snooze:{ts}`) is no longer expired an hour
  before it fires; also cancel a pending snooze when the plan is
  confirmed/skipped. Extract the sweep into a tested `expire_stale_nudges`.
- meridian-core/notifications: cache `has_action_columns` per pool, latching
  true only (the 057 columns are monotonic; a transient probe error stays
  false so the response leg is never disabled), so the pragma stops running
  on every tick forever. Add `.context(...)` to the two response-leg writes.
- notifications: add `NewNotification::interactive(category)` pairing
  category + actions from the one source of truth; move all three producers
  (daily_plan, worklog scheduler, notices) onto it so they can't drift.
- meridian-core/notifications: split the 888-line file (response leg →
  `responses.rs`, tests → `tests.rs`); all three are now under the 500 cap.
- tray: register notification categories synchronously before the poll loop
  starts, so a nudge already due at launch no longer renders button-less;
  collapse the duplicate bundle check (`update::is_packaged` now delegates to
  `sys::is_bundled`).
