+++ src/daily_plan.rs
@@ -33,20 +33,11 @@ pub async fn maybe_nudge(pool: &SqlitePool) -> Result<()> {
     let dedup = format!("plan.nudge:{today}");
     let now_utc = utc_iso(&now);
 
-    // Expire any still-live nudge from a previous day. Rows written before
-    // expiry stamping existed have `expires_at` NULL, so yesterday's unactioned
-    // banner would otherwise sit under today's identical one forever.
-    sqlx::query(
-        "UPDATE notifications SET expires_at = ?
-         WHERE event_key = 'plan.nudge' AND dedup_key <> ?
-           AND (expires_at IS NULL OR expires_at > ?)",
-    )
-    .bind(&now_utc)
-    .bind(&dedup)
-    .bind(&now_utc)
-    .execute(pool)
-    .await
-    .context("expiring stale plan nudges")?;
+    // Expire any still-live nudge that isn't part of *today's* nudge — a
+    // previous day's unactioned banner would otherwise sit under today's
+    // identical one forever. Rows written before expiry stamping existed have
+    // `expires_at` NULL, so they're swept here too.
+    expire_stale_nudges(pool, &dedup, &now_utc).await?;
 
     // Already confirmed or skipped today? Expire today's nudge (its job is
     // done — don't make the user dismiss it) and stop.
@@ -59,12 +50,15 @@ pub async fn maybe_nudge(pool: &SqlitePool) -> Result<()> {
         let confirmed: Option<String> = row.try_get("confirmed_at").unwrap_or(None);
         let skipped: i64 = row.try_get("skipped").unwrap_or(0);
         if confirmed.is_some() || skipped != 0 {
+            // Prefix match (not exact) so a pending snooze re-enqueue
+            // (`plan.nudge:{today}:snooze:{ts}`) is cancelled too — once the plan
+            // is actioned, a deferred reminder for it is moot.
             sqlx::query(
                 "UPDATE notifications SET expires_at = ?
-                 WHERE dedup_key = ? AND (expires_at IS NULL OR expires_at > ?)",
+                 WHERE dedup_key LIKE ? AND (expires_at IS NULL OR expires_at > ?)",
             )
             .bind(&now_utc)
-            .bind(&dedup)
+            .bind(format!("{dedup}%"))
             .bind(&now_utc)
             .execute(pool)
             .await
@@ -97,8 +91,7 @@ pub async fn maybe_nudge(pool: &SqlitePool) -> Result<()> {
         "Pick what you're working on today so Meridian can match your work to the right tickets.",
     )
     .link("/plan")
-    .category(categories::PLAN_NUDGE)
-    .actions(categories::actions_json(categories::PLAN_NUDGE).unwrap_or("[]"));
+    .interactive(categories::PLAN_NUDGE);
     if let Some(exp) = expires.as_deref() {
         nudge = nudge.expiring(exp);
     }
@@ -122,9 +115,83 @@ fn next_local_midnight_utc(now: &chrono::DateTime<Local>) -> Option<String> {
     Some(utc_iso(&local))
 }
 
+/// Expire every live `plan.nudge` row whose dedup key isn't part of *today's*
+/// nudge. Match on today's `plan.nudge:{today}` **prefix**, not exact equality:
+/// a snoozed nudge re-enqueues under `plan.nudge:{today}:snooze:{ts}` with
+/// `expires_at` NULL, so an exact `<>` check would sweep the snooze row and
+/// stamp its `expires_at` an hour *before* its `scheduled_for` — silently
+/// killing the deferred reminder before it ever fires. The prefix keeps today's
+/// canonical row AND its snooze re-enqueues alive while still expiring prior
+/// days. `today_dedup` carries no LIKE metacharacters (`%`/`_`), so no `ESCAPE`
+/// clause is needed.
+async fn expire_stale_nudges(pool: &SqlitePool, today_dedup: &str, now_utc: &str) -> Result<()> {
+    sqlx::query(
+        "UPDATE notifications SET expires_at = ?
+         WHERE event_key = 'plan.nudge' AND dedup_key NOT LIKE ?
+           AND (expires_at IS NULL OR expires_at > ?)",
+    )
+    .bind(now_utc)
+    .bind(format!("{today_dedup}%"))
+    .bind(now_utc)
+    .execute(pool)
+    .await
+    .context("expiring stale plan nudges")?;
+    Ok(())
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[tokio::test]
+    async fn stale_sweep_spares_today_and_snooze_but_expires_prior_days() {
+        use sqlx::sqlite::SqlitePoolOptions;
+        let pool = SqlitePoolOptions::new()
+            .max_connections(1)
+            .connect("sqlite::memory:")
+            .await
+            .unwrap();
+        sqlx::query(
+            "CREATE TABLE notifications (id INTEGER PRIMARY KEY, event_key TEXT, \
+             dedup_key TEXT, expires_at TEXT)",
+        )
+        .execute(&pool)
+        .await
+        .unwrap();
+        // id1: today's canonical nudge.  id2: today's snooze re-enqueue.
+        // id3: a prior day's unactioned nudge.
+        for (id, key) in [
+            (1, "plan.nudge:2026-07-07"),
+            (2, "plan.nudge:2026-07-07:snooze:2026-07-07T09:00:00Z"),
+            (3, "plan.nudge:2026-07-06"),
+        ] {
+            sqlx::query(
+                "INSERT INTO notifications (id, event_key, dedup_key) \
+                 VALUES (?, 'plan.nudge', ?)",
+            )
+            .bind(id)
+            .bind(key)
+            .execute(&pool)
+            .await
+            .unwrap();
+        }
+        let now = "2026-07-07T10:00:00Z";
+        expire_stale_nudges(&pool, "plan.nudge:2026-07-07", now)
+            .await
+            .unwrap();
+        let rows: Vec<(i64, Option<String>)> =
+            sqlx::query_as("SELECT id, expires_at FROM notifications ORDER BY id")
+                .fetch_all(&pool)
+                .await
+                .unwrap();
+        assert_eq!(rows[0].1, None, "today's canonical nudge must survive");
+        assert_eq!(rows[1].1, None, "today's snooze re-enqueue must survive");
+        assert_eq!(
+            rows[2].1.as_deref(),
+            Some(now),
+            "a prior-day nudge must be expired"
+        );
+    }
 
     #[test]
     fn expiry_is_the_next_local_midnight() {
